Use views for page layout
