+++ package.json
@@ -46,6 +46,7 @@
     "lodash": "^4.13.1",
     "morgan": "^1.7.0",
     "node-sass": "^3.8.0",
+    "pug": "^2.0.0-beta3",
     "react": "^15.0.1",
     "react-dom": "^15.0.1",
     "react-redux": "^4.4.5",
