+++ server/index.js
@@ -1,20 +1,5 @@
-import express from 'express';
-import morgan from 'morgan';
-import webpackMiddleware from './webpack';
+import app from './app';
 import logger from './logger';
-import router from './router';
-
-const app = express();
-
-if (process.env.NODE_ENV === 'production') {
-  app.use(express.static('dist'));
-} else {
-  app.use(webpackMiddleware);
-}
-
-app.use(morgan('dev'));
-app.use(express.static('assets'));
-app.use(router);
 
 const port = process.env.PORT || 8080;
 app.listen(port, error => {
