+++ server/views/index.pug
@@ -1,0 +1,10 @@
+doctype html
+html
+  head
+    title React London
+    link(rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/4.1.1/normalize.min.css")
+    link(rel="stylesheet" type="text/css" href="main.css")
+  body
+    main(id="main")
+      != content
+    script(src="/static/editor.js")
