+++ server/webpack.js
@@ -2,11 +2,14 @@ import webpack from 'webpack';
 import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackConfig from '../webpack.config';
 
-const compiler = webpack(webpackConfig);
-const middleware = webpackDevMiddleware(compiler, {
-  noInfo: true,
-  stats: 'errors-only',
-  publicPath: webpackConfig.output.publicPath,
-});
+function createWebpackMiddleware() {
+  const compiler = webpack(webpackConfig);
+  const middleware = webpackDevMiddleware(compiler, {
+    noInfo: true,
+    stats: 'errors-only',
+    publicPath: webpackConfig.output.publicPath,
+  });
+  return middleware;
+}
 
-export default middleware;
+export default createWebpackMiddleware;
